requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.2
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
looker-sdk==25.2.0
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.65.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
